Renamed module to the original sample name (excluding "in compose")
