+++ edit-and-sync-features-with-feature-service/src/main/AndroidManifest.xml
