+++ edit-and-sync-features-with-feature-service/src/main/java/com/esri/arcgismaps/sample/editandsyncfeatureswithfeatureservice/MainActivity.kt
@@ -14,7 +14,7 @@
  *
  */
 
-package com.esri.arcgismaps.sample.editandsyncfeatureswithfeatureserviceincompose
+package com.esri.arcgismaps.sample.editandsyncfeatureswithfeatureservice
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -25,7 +25,7 @@ import androidx.compose.runtime.Composable
 import com.arcgismaps.ApiKey
 import com.arcgismaps.ArcGISEnvironment
 import com.esri.arcgismaps.sample.sampleslib.theme.SampleAppTheme
-import com.esri.arcgismaps.sample.editandsyncfeatureswithfeatureserviceincompose.screens.MainScreen
+import com.esri.arcgismaps.sample.editandsyncfeatureswithfeatureservice.screens.MainScreen
 
 class MainActivity : ComponentActivity() {
