+++ edit-and-sync-features-with-feature-service/src/main/res/drawable/ic_launcher_background.xml
