+++ edit-and-sync-features-with-feature-service/src/main/res/mipmap-anydpi-v26/ic_launcher.xml
