+++ edit-and-sync-features-with-feature-service/src/main/res/mipmap-hdpi/ic_launcher.png
