+++ edit-and-sync-features-with-feature-service/src/main/res/mipmap-mdpi/ic_launcher_round.png
