+++ edit-and-sync-features-with-feature-service/src/main/res/mipmap-xhdpi/ic_launcher.png
