+++ edit-and-sync-features-with-feature-service/src/main/res/mipmap-xxhdpi/ic_launcher.png
