+++ edit-and-sync-features-with-feature-service/src/main/res/mipmap-xxxhdpi/ic_launcher_round.png
