+++ edit-and-sync-features-with-feature-service/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">Edit and Sync Features with Feature Service in Compose</string>
+    <string name="app_name">Edit and Sync Features with Feature Service</string>
     <string name="feature_server_url">https://sampleserver6.arcgisonline.com/arcgis/rest/services/Sync/WildfireSync/FeatureServer</string>
     <string name="dialog_title">Fetching result</string>
     <string name="generate_button_text">Generate Geodatabase</string>
